backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]
sqlalchemy
pydantic-settings
playwright==1.41.2
beautifulsoup4
readability-lxml
requests
openai>=1.30.0
google-generativeai>=0.7.0
typing-extensions>=4.10.0
python-multipart
lxml_html_clean
apscheduler
aiohttp
duckduckgo-search
pytz
tzdata
markdown
